moved date formatting into module
