+++ server/routes/reflection.js
@@ -18,6 +18,7 @@ var Registration = require('../models/registration');
 
 //our modules
 var generateSessionObject = require('../modules/sessionObject');
+var convertCount = require('../modules/convertCount');
 
 ///get reflections from database
 router.get('/', function (req, res) {
@@ -33,7 +34,7 @@ router.get('/', function (req, res) {
 
 router.get('/countByDay', function(req, res){
   console.log('count by day hit');
-  var reflectionCountByDate = [];
+
   Reflection.aggregate(
     [{$group: {
       _id : { month: {$month : '$reflectionDate'}, day: {$dayOfMonth: '$reflectionDate'}, year: { $year : '$reflectionDate'}},
@@ -43,16 +44,7 @@ router.get('/countByDay', function(req, res){
     if (err){
       console.log('error in count by day: ', err);
     }
-    for (i=0; i<countData.length; i++){
-      var countByDate = {
-        date: '',
-        count: 0
-      };
-      countByDate.date = countData[i]._id.month + '/' + countData[i]._id.day + '/' + countData[i]._id.year;
-      countByDate.count = countData[i].count;
-      reflectionCountByDate.push(countByDate);
-    }
-    console.log('countByDate: ', reflectionCountByDate);
+    var reflectionCountByDate = convertCount(countData);
     res.send(reflectionCountByDate);
   });
 });
